+++ Assets/CostCenter/Runtime/CCConstant.cs
@@ -3,5 +3,6 @@ namespace CostCenter {
     {
         private const string PLAYERPREFS_PREFIX = "CostCenter";
         internal static readonly string FIRST_OPEN_KEY = $"{PLAYERPREFS_PREFIX}_FirstOpen";
+        internal static readonly string TRACKED_ATT_KEY = $"{PLAYERPREFS_PREFIX}_TrackedATT";
     }
 }