+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.1.0
 - Add tracking Android install referrer
 - Add tracking iOS attribution token
+- Add tracking ATT
 
 ## v1.0.0
 - Attribution tracking